+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Development
 -----------
 
-* Placeholder
+* Add function to clean billing data to fit caltrack specifications (`clean_caltrack_billing_data`).
 
 2.7.6
 -----